ui/requirements.txt
click==8.1.3
Flask==2.1.2
flask-swagger-ui==3.36.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
Werkzeug==2.1.2
